temp
temp

+++ Source/Portfolio/Private/UI/Player_HPBar.cpp
@@ -6,7 +6,7 @@
 #include "Component/CharacterComponent.h"
 #include "Character/PlayerCharacter.h"
 
-void UPlayer_HPBar::SetMaxHp(float _MaxHp)
+void UPlayer_HPBar::SetMaxHp(int _MaxHp)
 {
 	SetMaxFigure(_MaxHp);
 }
@@ -16,14 +16,14 @@ void UPlayer_HPBar::InitalizeHPBarWidget(UCharacterComponent* NewComponent)
 	OnCurrentHpChange(0, NewComponent->GetCurrentHp());
 }
 
-void UPlayer_HPBar::OnMaxHpChange(float _MaxHp, float NewMaxHp)
+void UPlayer_HPBar::OnMaxHpChange(int _MaxHp, int NewMaxHp)
 {
 	SetMaxFigure(NewMaxHp);
 
 	OnCurrentHpChange(0, NewMaxHp);
 }
 
-void UPlayer_HPBar::OnCurrentHpChange(float _Hp, float NewHp)
+void UPlayer_HPBar::OnCurrentHpChange(int _Hp, int NewHp)
 {
 	if (IsValid(Bar)) {
 		if (MaxFigure > KINDA_SMALL_NUMBER) {
